remove -X from vim command
The -X flag to vim is now controlled by setting "let
vimpager_disable_x11 = 1" in .vimrc or .vimpagerrc.

Reorganize the code a bit so that the vimpager variables can be set in
.vimpagerrc / VIMPAGER_RC as well, not just ~/.vimrc.

Update vimpager.md to note the new variable and rebuild man page and
README.

This hopefully fixes (GitHub issue #19).