fix "col: illegal byte sequence" with git log -p
col -b errors out on byte sequences it doesn't like, cutting off the
output, as seen in git log -p output (GitHub Issue #24).

Using sed doesn't always work because the input can be utf8.

So we check for perl 5.8.1, and if available, use perl -CIOL, otherwise
sed.

It's very rare that the input is utf8 and sed doesn't work for this or
whatever other reason, so this should be the best way of doing it.