noremap up/down mappings into preceding -c
Squash the new up/down mappings into the preceding -c with the
:nohlsearch mapping to keep the number of -c commands low, because there
is a limit on them.

Change the nmap to an nnoremap, as that seems to work much better. For
me with just an nmap <Down> was jumping too many lines down.

Add the mapping to the GUI vim invocation section as well.

Add iftekhar25 to CONTRIBUTORS section.