stop sourcing vimrc/gvimrc twice
Sourcing vimrc/gvimrc twice was causing problems for some people:

https://github.com/rkitover/vimpager/pull/32

however this was done in 7b92857 because the "syntax on" statement in
less.vim resets highlights.

The requestor suggested a more elegant solution to that issue, which
also fixes his, that is to use --cmd 'runtime! macros/less.vim' instead
of -c to source less.vim. This seems to not cause any problems, and we
can remove the re-sourcing of the rc files.