major refactor and experimental vimcat utility
Fix a bug in de3d11c where filename arguments would cause vimpager to
just hang.

Also make de3d11c turn on only if the vimpager_passthrough option is
set. The rationale for this is we cannot yet do highlighting for text
dumped to the terminal. The vimcat utility (based on _v by Magnus
Woldrich: https://github.com/trapd00r/utils/blob/master/_v) almost does
this, but has some highlighting bugs on OSX.

When the vimcat utility is functional, we'll have to turn vimpager into
a dist and install both of them, with vimpager using vimcat, but that's
when it works.

Also turn off the passthrough option if the shell doesn't have
arithmetic expansion, as that code relies on it.

Major refactor of option reading, vimrc/gvimrc finding, and tempdir
handling for more portability.

Makes the code generally less ugly.