fix usage with no .vimrc (Github Issue #38)
Now that every vim/gvim invocation uses -u to point to the vimrc, which
is read from vim's $MYVIMRC variable, there is a problem if there is no
user vimrc ($MYVIMRC is empty) and -u reads the next argument as the
vimrc, completely breaking vimpager.

Detect this situation and set vimrc to 'NORC'.