nocp on start, fixes pathogen (Github Issue #40)
Now that we always use -u to load the user's vimrc, we have to deal with
a sideeffect of using -u: namely that it does a set compatible.

Fix this by adding set nocp to the --cmd list of commands everywhere vim
-u is used.

Pathogen aborts loading when &cp is set, this is why it did not work
with vimpager.