fix screen flashing/misc. fixes (GitHub Issue #46)
vim only reads the .vimrc when input is from a terminal, so there is no
way to read $MYVIMRC without the screen flashing.

Factor out the .gvimrc finding into a function find_rc() and use it to
also find the .vimrc, using the heuristics described in :help startup.

Other improvements in this commit:

 * quote all files/dirs (especially protect ${tmp} with spaces)

 * add krijesta and vincer to CONTRIBUTORS