don't clobber user's PATH #77
Since PATH is usually exported, reexecing a shell will preserve the PATH
fixup for finding a shell, which has the existing PATH at the end.

Move $PATH to the beginning, so that the user's PATH overrides the extra
directories we are searching for a shell, and is preserved on reexec so
that the correct vim and other binaries are launched.

I did think this was wrong when I wrote the code initially, but I didn't
think about the impact of previously exported env vars in a script.