bundle AnsiEsc and redo logic of #81 (WIP)
We now bundle the improved version of AnsiEsc
(http://www.vim.org/scripts/script.php?script_id=4979) as a uuencoded
tar inside the script which expands at runtime into a prepended temp
directory in runtimepath.

AnsiEsc is only activated if the "conceal" feature is available, it is
not on the default vim on OS X, but is in the Homebrew version of vim.

Also there is no longer an option to turn on AnsiEsc, it gets turned on
automatically if the source file has ANSI codes and the conceal feature
is available.

Document all of this and rebuild docs.

Also move the ~/.bashrc settings into the INSTALL section.

Rebuild docs.

A copy of the above mentioned AnsiEsc has been placed into the ansiesc/
directory inside the project, and it gets automatically installed into
the vimpager script when you run "make".

The uuencoded tar is not gzipped, in case the system does not have
gunzip (e.g. bare Solaris installs.) This makes the script much larger,
but the function is at the end and will not get too much in the way of
development.

TODO:

 * whitespace errors bleed in git diff, this needs fixing in AnsiEsc

 * find a version of uudecode implemented in awk, so as not to rely on
   the system version, which may not be available.