fix compatibility with GNU sharutils uudecode #87
GNU sharutils uudecode does not support the -p option for outputting to
STDOUT, but it does allow '-o -', which BSD uudecode takes to mean a
file called literally '-'.

Fortunately, GNU uudecode is documented to support '-o /dev/stdout' and
that works on BSD uudecode as well, so switch to that syntax.