don't drop last line in passthrough mode #102
In the passthrough mode calculation loop, we were checking the return
value of read to see if a line was read.

When read reads a trailing line that has characters but no newline and
hits an EOF, it still reads a line but returns false.

Adjust the code to check BOTH the return value of read and whether the
line was read, this way trailing lines without a newline are also read
in this loop.

Thanks to geirha on freenode ##bash for explaining how this works to me.