space/w ansi, passthru on, utf8/win32 #103 #11 #104
Fix the horrible bug that produced incorrect spacing when using AnsiEsc,
finally.

Bundle vimcat in vimpager and use it for the passthrough mode when there
are no ansi escapes or overstrikes.

Passthrough mode now defaults to on. Update docs.

Don't use perl to remove overstrikes on cygwin/msys because it may be a
Win32 native perl and may have issues detecting the locale.