fix passthrough mode #11
Passthrough mode now preserves extra vim commands, like enabling man
page highlighting for small man pages and uses the filename when
available to choose a syntax as well.

Multiple -c and --cmd options are now supported in both vimpager and
vimcat, as well as the -u option for the .vimrc.