remove generated files from git (merge #113)
The README file and the man pages can be made by the makefile.  Removing
them from git will ensure that users will make a current version from
the source before installing them.

The file README.md is turned into a symlink in order to be usable for
the github project page.