make target install depend only on docs
Make the install makefile target depend only on docs, which displays a
warning if pandoc is not installed, and not on all which also rebuilds
vimpager. This will help if the user does not have uuencode on the
system.

I'm trying currently to rewrite the Makefile to only incrementally
repackage vimpager if any of the bundled components have changed, then I
will revert to depending on all for the install target.