more Solaris fixes
Add install-sh from automake for a portable replacement for install for
make install, this also helps on Solaris because the Solaris install
works differently and you have to use /usr/ucb/install for a more
standard version.

Remove the ?= construct from the Makefile variables. It is not necessary
because according to POSIX variables specified on the make command line
override those set with a regular = in the Makefile, and the ?=
construct is not supported by Solaris make.

Add an AWK Makefile variable, that defaults to awk.

Replace the one line lookup table for the awk uudecode with separate
lines each no longer than 2559 characters, as this is the limit for
/bin/awk on Solaris, and longer lines break the build as awk is used to
assemble the source files into uuencoded versions inside vimpager. This
involved modifying gen_awk_uudecode_lookup.py to do the line splitting
and replacing the generated code in vimpager.

Detect if /usr/xpg4/bin/tail is available in vimcat when checking tail
-c 1, because Solaris /bin/tail does not have the -c option.

Issue a cd ${tmp}/.. before removing ${tmp} as Solaris does not allow
removing the current directory.