turn hlsearch back on #121
Make searches highlight by default. I had previously thoughtlessly
disabled it because I usually have it off, but it makes more sense for
it to be on by default with a key to toggle it.

Make <ESC>u actually toggle the highlighting, rather than just turn it
off. This doesn't always work, sometimes you have to press it twice,
I'll try to figure out why later.