make Makefile fully incremental
The Makefile now does a real incremental build of vimpager and the man
pages. Works perfectly on Solaris make as well.

uuencode is no longer necessary for building an up to date repo.

AnsiEsc is now stored in vimpager as the three .vim files, not as a tar,
this makes the incremental build simpler.

The output of the make process is now much more concise and
user-friendly. Just one line messages instead of huge screenfulls of
cryptic commands.