vimcat: visual as first --cmd, fix :endifs
Enter visual mode in vimcat as the very first --cmd, so that user --cmd
and -c commands are in visual mode.

Fix number of :endif commands after the :if 0 block enclosing the shell
script portion.

The source command for the vim script has to be run through silent!
because it complains about the first # shell comment in the :if 0 block,
giving the error "trailing characters".