add -o output_file option for vimcat
Add -o option to specify output file for vimcat, with this option
whether or not output is a terminal is ignored. If output_file is - then
the output is written to STDOUT, but without the extra newline if there
isn't one on the last line.

If there is no -o option, then the normal behavior of passing through to
cat if the output is not a terminal is preserved.

Update docs/vimpager.md to describe the -o option.