vimcat: go back to using matchstr not strpart
Go back to using matchstr to get the current column from a line instead
of strpart because matchstr works on characters while strpart works only
on bytes, and this can break unicode text.

Also I did not use strpart correctly anyway, it uses a zero index but
the column index is one based, so it was eating the first character of
lines.

Also check for a concealed item earlier and don't output ANSI codes for
concealed items.