update CONTRIBUTORS, docs and LICENSE inf
Update LICENSE to specify which parts of the project are BSD 2-clause
and which parts are Vim license, and state the dual-license nature of
the project more explicitly.

Include the Vim license, uganda.txt as required by the Vim license and
refer to it in LICENSE.

Add willywampa to CONTRIBUTORS for b45896e .

Update date stamps on docs.

Add a note to doc/vimcat.md about the bg=dark default and how to
override it, and give more details about the "-o -" option.