vimcat: set bg color to NONE by default
Many color schemes set the ctermbg (terminal background color) and this
does not work well in vimcat, because it outputs ANSI reset codes at the
end of every line.

TODO: fix this so that only necessary reset codes are output

Add the command:

  hi Normal ctermbg=NONE

to the list of default commands for vimcat.

I would suspect most people don't want the background color even if it
worked correctly, although fixing this would still be a good thing.

This can be overridden with a -c option by e.g. re-running the colors
command, like -c 'colors apprentice' .

Rebuild vimpager with updated vimcat.