fix ANSI detection regex (again) #125
Fix ANSI detection regex to not always match (non-ANSI files as well.)

Change the * on the outer () group to a +, so that at least one ANSI
code must be matched.

This behavior is triggered in Ubuntu gawk 4.1.1, for whatever reason it
was not triggered on my Mac which I believe was using gawk 4.1.3. Will
need to investigate further.