add vimpager -N option for line numbers #136
Add -N or --LINE-NUMBERS option which sets g:less.number = 1 which does
a set nu in less.vim . By default set nonu is run instead, regardless of
the user's nu setting for this.

This is done by passing a dict of opts to vimpager#Init(opts), other
options can easily be later added to this.