temporarily set modifiable in ConcealRetab() #137
In case the user has `nomodifiable` set for the buffer and is using
AnsiEsc, temporarily setlocal modifiable during the tab fixup in
ConcealRetab() and restore it back to the user's setting. This is
necessary because the function changes the number of spaces in the file
to get the tabs to line up correctly.

Move the plugin blacklist into DisableConflictingPlugins().

Use `silent! args` instead of `silent! e` to read the file, in
preparation for multiple file support.