multiple file support #119
Support opening multiple files, including files with ANSI markup.

Write the necessary vim commands to ${file}.vim in $tmp files which are
executed on `BufWinEnter`.

man and perldoc support is disabled when multiple file arguments are
given, with the exception of files where the NAME regex matches, those
will still be `set ft=man`.

Add `:n` and `:N` to docs and help.

Check if file arguments are readable and exit with an error if not.

Add a quit function that cleans up the temp dir instead of doing this at
every exit point.