display actual file name on the status line
Add the config key `g:less.statusfunc` which is a funcref, this is
overridden in vimpager.vim to remove the temp directory and urldecode
the file path, which is sent urlencoded from vimpager.

The default one left in less.vim does almost the same thing, without the
temp dir removal and urldecoding.