preliminary support for neovim in vimcat #133
Remove the -X option from the vim invocation in vimcat because neovim
does not support it. The extra few milliseconds saved by not connecting
to the X server is not worth doing a version check on the vim
executable.

Before invoking vim, check that either vim or nvim, in that order, are
in the PATH and use the first one available.

If neither vim nor nvim are available in the PATH, throw an error.