rework vimrc finding mechanism #141
Read the location of the user's .vimrc by reading the $MYVIMRC variable
from vim. Check for vimpagerrc in additional places as described in :h
initialization.

Remove the check for gvimrc, the color resetting hack may not be
necessary anymore, will need to check, and the simalt ~r is definitely
unnecessary. The column adjustment will move into vimpager.vim, for now
just set columns.