disable viminfo #141
Disable viminfo with -i NONE, this will avoid writing a viminfo file to
possibly an unintended path, and writing entries to viminfo for
non-repeatable temp files.

viminfo handling will need to be improved in the future, it would be
nice if the user's position in a man page or other doc was saved and
restored for example.