use markdown_github pandoc format for html
This formt highlights at least the bash code snippets, but not the vim
code snippets yet. I might try to add support to the relevant Haskell
modules for this.

TODO: Not using this format to generate man pages yet because it does
not understand the top man metadata, or perhaps it needs to be in a
different format.