sep. shell detect, stream vimcat out #149 #152
Move POSIX shell detection into inc/prologue.sh, which is sourced by
the repo version of the scripts, and included in the standalone
versions. In the installed versions it is neither sourced nor included,
but the shebang is changed to the reflect the shell found at install
time, as discussed in #149.

A couple of changes from PR #152 are merged manually because it could
not be merged automatically, thanks @lucc.

Make vimcat use a fifo read with a background tail -f, to which the vim
script writes lines in append mode. This should make for smoother
scrolling output from vimcat. Input buffering will be done in the
future as well.