Rework file preprocessing in the shell script.
Files are not copied to /tmp unconditionally but only if they are modified
(e.g. decompressed, ansi escape sequences removed, ...).  If input is read
from stdin it is still saved to /tmp and paged from there.

The file specific vim scripts are written to numbered files in /tmp (according
to the position of the original file in the argument array).  This makes them
agnostic about the fact that the original file might be in /tmp or somewhere
else.