fix incompatibility with dein#load_state() #166
Disable the loading of all of our plugins at --cmd time to make sure
older versions the user may have are not loaded, and then explicitly
re-add the vimpager runtime to the runtimepath at BufReadPre time and
explicitly load all plugins from it.

The main problem with dein#load_state() is that it completely overwrite
the runtimepath, hence re-adding it is necessary.

Older versions of plugins were also in the reporter's vim configuration,
so this commit solves both problems.