add input streaming for vimcat #172
Implement a pipeline with split -b 4096 as the endpoint into a temp dir
for vimcat, with the viml reading chunks as needed and deleting from
the top to keep a window onto the stream.

This is a prototype for how vimpager will implement this as well,
although that will be a bit more complicated.

Other misc improvements:
- add VIMCAT_DEBUG env var to not silence vim when sourcing script and
  running highlight code
- allow '-' file specifier for stdin for vimcat and vimpager
- call quit 1 on trap in vimcat and vimpager
- display error if -o is used for more than one file
- use the same method to separate files in vimcat as in vimpager, with
  blank lines between the headers and files
- truncate output file with -o option
- update ChangeLogs and TODO

TODO:
- ./vimcat -o - file | head # hangs the terminal, need to figure out
  why and fix