remove redundant cmds from .travis.yml, fix build
Remove the apt-get update command from .travis.yml because make
install-deb does this and it was breaking the build because of a hash
sum mismatch.

Remove the apt-get install lintian command because update_lintian builds
the newest one anyway.

Add || true to the apt-get update commands in update_lintian just in
case it may be used +e.