replace grep_q with grep -Eq, check for best grep
POSIX egrep is grep -E, which is the same regular expression syntax as
awk, which we were using from the grep_q function.

Solaris actually suports -q in /usr/xpg4/bin/grep.

Replace all grep_q calls with grep -Eq invocations.

Check for ggrep, Solaris /usr/xpg4/bin/grep and then fall back to grep.

Check that the grep -Eq syntax is supported and works.

In the grep function override, if -Eq is the only option and is not
supported, call the previous implementation of grep_q renamed to
awk_grep_E_q().