a couple of fixes for neovim #67
Set laststatus=1 in neovim, which is the vim default, so that the
interface is consistent and there is no status line by default.

Fix the q mapping in less.vim to adjust for differences in :ls in
neovim, our buffers end up unlisted so use :ls! instead of :ls, and
adjust the regex for checking for only one buffer for differences in
flags in neovim.