Split the building process for standalone/vimpager.
The files that need to be bundled with vimpager are uudecoded, the result
wrapped into the correct shell function and that saved to a new file with a
`.uu` suffix.  This allows to clean up inc/bundled_scripts.sh and simplifies
the insertion of the uudecoded data into the final script.  Additionally the
dependencies on the bundled files are cleanly stated in the makefile and only
updated files need to be uudecoded again.