Split vimcat file.
Most vim code is moved to autoload/vimcat.vim.  A new configuration variable
is added at the top of the script to hand the extra runtimepath component to
find this file to vim.  Like this the vimcat script does not need to source
itself any more.

The balancing statements are removed.  The shell code is reindented to match
the new modeline.  The main `set - ...` command before calling vim is
reformatted to make it more readable and to make it easier to modify it from
the makefile.

Some superfluous comments and quotes are removed.