ver tag var at top of progs/buid from src tarball
Set the latest version tag as version_tag=XXX at the top of vimpager and
vimcat.

If version assignment fails due to not having .git or a shallow clone,
use the version_tag var with the postfix '(checkout)'.

On make for the all recipe, update the version_tag var from either git
or the ChangeLog files.

For make recipes that used `git describe`, write out a XXX-version.txt
file instead with the output of `git describe`, or the latest tag from
ChangeLog if `git describe` has failed.

This allows building from a source tarball (rather than only a git
clone.)