viml initialization improvements
Move the execution of post-processing .vim files written by vimpager
shell code into a function that runs only once if hidden is set, and on
every BufWinEnter otherwise.

Save the indices of the .vim files at initialization because the 'q'
mapping modifies the argument list and argidx() cannot be relied upon.

Change the function that turns on AnsiEsc to ignore the filetype set
from a modeline if the file looks like a diff (either git diff or
unified colordiff.) This is done so that modelines that happen to show
up in diffs do not break the colorization.

TODO: match normal (non-unified) diffs as well.

Change less.vim to enable less mode if g:less.enabled is set on
BufWinEnter instead of BufReadPost. This unifies the code into a single
EnterWindow() function and gives LessMode() a better change of
overriding conflicting mappings set by filetypes, for example the neovim
man.vim overrides 'q' which we need in less.vim, with this change
LessMode() runs later and overrides the man.vim mapping.