plugin and :Page command improvements
Set global keywordprg to `:Page!\ -t\ man` if it's set to `man ...` or
`:Man`.

Add FileType autocommands for python, ruby, perl and sh to set
buffer-local keywordprg for the appropriate documentation command.

TODO: add a mapping for ft=vim with :help, and add support for
`:Page!  :vim-cmd`.

Fix bug in Page that turned on less mode for the buffer where the
command was invoked, while preserving the value of `g:less.enabled`.

For `:Page!` commands, redirect STDERR to /dev/null .

Run overstrike/ANSI substitution commands with silent so that no prompts
are generated.

For `:Page!` commands, setlocal buftype=nowrite, as well as for any
`:Page` commands where ANSI escapes are removed.