improve rc selection + default hl opts #203
Only pass -u $vimrc to vim in vimpager if necessary (a vimpagerrc exists
etc.) Otherwise let it use the defaults. I couldn't do this before
because I didn't know how to use the ${:+} POSIX parameter expansion
properly, it is amazingly useful.

Move the bg=dark setting for the case of no rc to after --cmd, because
$MYVIMRC is not set until later.

Add similar bg=dark setting code to the vimpager vimscript.