fix vim-plug compat #221
vim-plug expects its plugin store directory to be in the first entry in
runtimepath, which is generally ~/.vim or similar, because vimpager
prepends its runtime to the runtimepath, this fails.

Restore runtimepath to the default for the duration of ~/.vimrc
processing so that vim-plug and similar things work.

The vimpager runtime is added back to the runtimepath in BufReadPre for
any file so that our plugins still work correctly.