vimcat: use fifo not file, even on cygwin etc.
In 09fd981 the tail -f call for reading the fifo was replaced with a cat
loop, this works fine on cygwin and presumably everywhere else it didn't
work before, so remove the OS check for making a regular file instead of
a fifo and always make a fifo instead.

This also fixes vimcat on cygwin/msys since 09fd981 broke it, the cat
loop would repeat the file over and over.