fix terminal size detection without tput
Use `bash -O checkwinsize -i -c 'echo $LINES` and `$COLUMNS` to detect
the terminal size if tput is not available. The old code would set only
`$LINES` and not `$COLUMNS` and force pass-through for all files, no
matter how long.

If neither the `tput` nor the `bash` method for getting the terminal
size works, force pass-through and display a warning that `tput` is not
available.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>